requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx==0.27.0
cachetools==5.3.3